+++ runtime/std/streams-plugin/src/main/scala/scalascript/compiler/plugin/streams/StreamsIntrinsics.scala
@@ -697,7 +697,7 @@ object StreamsIntrinsics:
               var step = ctx.invokeCallback(nextFn, Nil).asInstanceOf[Value]
               while step != Value.NoneV do
                 step match
-                  case Value.OptionV(Some(v)) => queue.put(Some(v))
+                  case Value.OptionV(v) if v != null => queue.put(Some(v))
                   case _ =>
                 step = ctx.invokeCallback(nextFn, Nil).asInstanceOf[Value]
             catch case _: Throwable => ()
@@ -1216,7 +1216,7 @@ object StreamsIntrinsics:
                     val result = ctx.synchronized { ctx.invokeCallback(f, List(state)).asInstanceOf[Value] }
                     result match
                       case Value.NoneV => running = false
-                      case Value.OptionV(Some(Value.TupleV(xs))) if xs.length == 2 =>
+                      case Value.OptionV(Value.TupleV(xs)) if xs.length == 2 =>
                         q.put(Some(xs(1)))
                         state = xs(0)
                       case _ => running = false